noticed during ticket#86 (unfinished) - gwidget.s lastX and lastY need to be integers - they are passed as int to a C lib call
git-svn-id: https://aeon.stsci.edu/ssb/svn/pyraf/trunk@858 05fa7843-0683-0410-9e38-dfee850dc5be

+++ lib/gkigcur.py
@@ -77,8 +77,8 @@ class Gcursor:
     def cursorOff(self):
         """Turn cross-hair cursor off"""
         self.gwidget.deactivateSWCursor()
-        self.gwidget.lastX = self.x/self.gwidget.width
-        self.gwidget.lastY = self.y/self.gwidget.height
+        self.gwidget.lastX = int(self.x/self.gwidget.width)
+        self.gwidget.lastY = int(self.y/self.gwidget.height)
 
     def bind(self):
